Add multiple and simplify clear background
Add capabilities and flexibility for plotting multiple datasets within one call of the function. Will plot contours of different colours for each dataset.

Add capability and flexibility for user-defined plotting of contour colour fill, contour lines, and displaying contour level values.

Simplify and improve the ability of plotting a clear background of the contours (which is actually not plotting the lowest level of the contours, using 'prune='lower''). No more user-required input, and makes background transparent and not white, which is vital for plotting multiple datasets.